build: move parse-changelog to separate file
