+++ devtools/scripts/set-version.js
@@ -3,106 +3,7 @@ const util = require('util');
 const exec = util.promisify(require('child_process').exec);
 const chalk = require('chalk');
 
-/**
- * Creates Changelog Object from CHANGELOG.md
- */
-const parseChangelog = async () => {
-    const POS = {
-        OUT: 'OUT',
-        VER: 'VER',
-        TYPE: 'TYPE',
-        ENTRY: 'ENTRY',
-    };
-    const changelog = [];
-
-    const changelogFile = 'CHANGELOG.md';
-
-    const exprVersion = /^## \[(?<version>[\d.]+)\]/u;
-    const exprTypeStart = /^### (?<type>[a-zA-Z]+)/u;
-    const exprTypeEntry = /^- (?<entry>.+)/u;
-
-    let cursor = {
-        pos: POS.OUT,
-        version: null,
-        type: null,
-    };
-    const changelogContent = await fs.readFile(changelogFile, { encoding: 'utf-8' });
-    for (const entry of changelogContent.split("\n")) {
-        /**
-         * Version
-         */
-        const match = entry.match(exprVersion);
-        if (match) {
-            if (cursor.version) {
-                cursor.version.types.push({
-                    ...cursor.type,
-                });
-
-                changelog.push({
-                    ...cursor.version,
-                    desc: cursor.version.desc.join("\n").trim(),
-                });
-            }
-
-            cursor.version = {
-                version: match.groups.version,
-                desc: [],
-                types: [],
-            };
-            cursor.pos = POS.VER;
-
-            continue;
-        }
-
-        /**
-         * Version Description
-         */
-        if (cursor.pos === POS.VER) {
-            const match = entry.match(exprTypeStart);
-
-            if (!match) {
-                cursor.version.desc.push(entry)
-            }
-        }
-
-        /**
-         * Change Type Title
-         */
-        if (cursor.pos !== POS.OUT) {
-            const match = entry.match(exprTypeStart);
-
-            if (match) {
-                if (cursor.pos === POS.ENTRY) {
-                    cursor.version.types.push({
-                        ...cursor.type,
-                    });
-                }
-
-                cursor.type = {
-                    title: match.groups.type,
-                    entries: [],
-                };
-                cursor.pos = POS.TYPE;
-
-                continue;
-            }
-        }
-
-        /**
-         * Change Type Entry
-         */
-        if (cursor.pos === POS.TYPE || cursor.pos === POS.ENTRY) {
-            const match = entry.match(exprTypeEntry);
-
-            if (match) {
-                cursor.type.entries.push(match.groups.entry);
-                cursor.pos = POS.ENTRY;
-            }
-        }
-    }
-
-    return changelog;
-}
+const parseChangelog = require('../lib/parse-changelog');
 
 /**
  * Update the version in package.json and package-lock.json
@@ -172,6 +73,7 @@ const createReleaseCommit = async (version) => {
         if (error.stderr.match(/tag .* already exists/um)) {
             throw new Error('Version tag is already exists. Please delete it or increase the version number.');
         }
+
         console.log(error);
     }
 };
@@ -186,7 +88,8 @@ const setVersion = async (opts) => {
     }
 
     // Get current version
-    const changelog = await parseChangelog();
+    const changelogContent = await fs.readFile('CHANGELOG.md', { encoding: 'utf-8' });
+    const changelog = await parseChangelog(changelogContent);
     if (changelog.length === 0) {
         throw new Error(`Changelog has no entries`);
     }
